// src/interface/server.enum.ts
export enum ServerURL {
  DoesNotWorkPerfumeNew = 'https://does-not-work.perfume.new',
  GitLab = 'https://gitlab.com',
  AppScntMe = 'http://app.scnt.me',
  OfflineScentronixCom = 'https://offline.scentronix.com',
}

export enum ServerPriority {
  Low = 1,
  Medium = 2,
  High = 3,
  Highest = 4,
}